+++ reddwarf/utils/clusterer/base.py
@@ -1,0 +1,34 @@
+from typing import Optional
+
+from reddwarf.sklearn.cluster import PolisKMeans
+from reddwarf.utils.clusterer.kmeans import find_optimal_k
+
+
+def run_clusterer(
+    clusterable_participants_df,
+    clusterer="kmeans",
+    force_group_count=None,
+    max_group_count=5,
+    **clusterer_kwargs,
+) -> Optional[PolisKMeans]:
+    match clusterer:
+        case "kmeans":
+            if force_group_count:
+                k_bounds = [force_group_count, force_group_count]
+            else:
+                k_bounds = [2, max_group_count]
+
+            _, _, kmeans = find_optimal_k(
+                projected_data=clusterable_participants_df,
+                k_bounds=k_bounds,
+                # Force polis strategy of initiating cluster centers. See: PolisKMeans.
+                init="polis",
+                **clusterer_kwargs,
+            )
+
+            return kmeans
+
+        case "hdbscan":
+            raise NotImplementedError("clusterer type hdbscan not yet implemented")
+        case _:
+            raise NotImplementedError("clusterer type unknown")
